allow SCC to fail more often on fips
